# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "A production-ready geospatial routing microservice. Supports self-hosted OSRM (default) with an automatic in-process A* + Yen's k-shortest-paths fallback, live GPS tracking, WebSocket event delivery, and a PostGIS-backed OSM road graph.",
        "title": "geo-service",
        "contact": {
            "name": "GitHub Issues",
            "url": "https://github.com/Amirbehzad11/geo-service/issues"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/gps/trip/{id}/location": {
            "get": {
                "description": "Returns the most recent GPS state for a trip, read from Redis. State is cached for 24 hours after the last update.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "gps"
                ],
                "summary": "Get latest trip location",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Trip ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Latest location state",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/geo-service_internal_response.Success"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_gps.LocationState"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid trip ID",
                        "schema": {
                            "$ref": "#/definitions/geo-service_internal_response.Failure"
                        }
                    },
                    "404": {
                        "description": "No location found for this trip",
                        "schema": {
                            "$ref": "#/definitions/geo-service_internal_response.Failure"
                        }
                    }
                }
            }
        },
        "/gps/update": {
            "post": {
                "description": "Accepts a raw GPS position for an active trip. Applies rate limiting, EMA smoothing (α=0.75), speed computation, and cross-track deviation detection against the planned route. Writes state to Redis and broadcasts events via Redis Pub/Sub.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "gps"
                ],
                "summary": "Process a GPS update",
                "parameters": [
                    {
                        "description": "GPS update payload",
                        "name": "update",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_gps.GPSUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Processed location state",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/geo-service_internal_response.Success"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_gps.LocationState"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Malformed JSON body",
                        "schema": {
                            "$ref": "#/definitions/geo-service_internal_response.Failure"
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/geo-service_internal_response.Failure"
                        }
                    },
                    "429": {
                        "description": "Rate limited — update received too soon",
                        "schema": {
                            "$ref": "#/definitions/geo-service_internal_response.Failure"
                        }
                    },
                    "500": {
                        "description": "Internal processing error",
                        "schema": {
                            "$ref": "#/definitions/geo-service_internal_response.Failure"
                        }
                    }
                }
            }
        },
...
        "/ws/shipments/nearby": {
            "get": {
                "description": "Receives lat/lng over WebSocket. type=passenger queries the Laravel shipment table; type=sender searches nearby passengers in Redis.",
                "tags": [
                    "websocket"
                ],
                "summary": "WebSocket - nearby shipments",
                "responses": {
                    "101": {
                        "description": "Switching Protocols - WebSocket upgrade successful"
                    },
                    "503": {
                        "description": "Shipment database is not configured",
                        "schema": {
                            "$ref": "#/definitions/geo-service_internal_response.Failure"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "geo-service_internal_response.ErrorDetail": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "geo-service_internal_response.Failure": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/geo-service_internal_response.ErrorDetail"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "geo-service_internal_response.Success": {
            "type": "object",
            "properties": {
                "data": {},
                "success": {
                    "type": "boolean"
                }
            }
        },
        "internal_gps.GPSUpdate": {
            "type": "object",
            "properties": {
                "lat": {
                    "type": "number"
                },
                "lng": {
                    "type": "number"
                },
                "timestamp": {
                    "description": "Unix seconds",
                    "type": "integer"
                },
                "trip_id": {
                    "type": "integer"
                }
            }
        },
        "internal_gps.LocationState": {
            "type": "object",
            "properties": {
                "deviation_km": {
                    "description": "distance from planned route",
                    "type": "number"
                },
                "lat": {
                    "type": "number"
                },
                "lng": {
                    "type": "number"
                },
                "speed_kmh": {
                    "type": "number"
                },
                "timestamp": {
                    "type": "integer"
                },
                "trip_id": {
                    "type": "integer"
                }
            }
        },
...
    }
}